feat: remove create-joined-user handler, add create-user-from-discord handler
