+++ libs/domains/src/user/application/commands/create-user-from-discord/create-user-from-discord.input.ts
@@ -1,6 +1,9 @@
 import { IsOptional, IsString, IsUUID } from 'class-validator';
 
-export class CreateJoinedUserInput {
+export class CreateUserFromDiscordInput {
+  @IsUUID()
+  id: string;
+
   @IsString()
   username: string;
