+++ libs/domains/src/user/application/events/user.event.providers.ts
@@ -1,3 +1,1 @@
-import { JoinedUserCreatedHandler } from './joined-user-created/joined-user-created.handler';
-
-export const USER_EVENT_PROVIDERS = [JoinedUserCreatedHandler];
+export const USER_EVENT_PROVIDERS = [];
